CSS Setup (#1)
* Adjusted Mock DB

Changed the array returned by the mock db, so that it more closely reflects the structure implied by the mockup design table (e.g. separate firstname and lastname fields)

* Separated CSS from Core Module Code

Moved the style object to another file, and added additional styling.

* Mapped CSS from Mockup to Organizer Page

Took the CSS from the site mockup, and applied it to the organizer page.
Flattened rgba color values into rgb.
Results are presented in a table, rather than in sorted divs, this shouldn't cause any difference in terms of functionality or visual representation.

* Added a taskbar component

Added the component referred to as "taskbar" in the mockup.

* CSS fixes

Added minor styling changes to the <body>, through a style tag.

* Separated Scrolling Sections

Restructured the 'Organizer' page, making it a bit more responsive, mainly by making scrolling independent to different sections.

* Selection Mechanism

Added the ability to select a row out of the activists list (and deselect it), and to simultaneously select all presented rows (and deselect them).