Added a 2-Step Login Screen
Login requires entering a phone number or an email of an activist from the mock DB, and then entering the temporary fixed login code 123456 (very original).
The only query that requires a login token validation is getActivists().
The login token is saved in a cookie, and in the mock DB. Seeing as the mock DB isn't persistent, re-logging in is required each time after refreshing.
Change line 21 in ItemsService if you are working on Organizer.js.