Added routing handler, Mongoose Queries
Routing is now handled automatically, by scanning the routes directory for files.

Added a "fetch activists" function, that interfaces with the db, and can be called via a get request to /api/activists.

Added an "insert activist" function, that interfaces with the db, and can be called via a post request to /api/activists.