Fixed unclear error with fetch function
when trying to make a request to the server from the Caller page, an error was produced - "fetch is not defined". I'm not sure why it only appeared from that page, but I fixed it by explicitly importing fetch frome node-fetch, and switching the url paths from relative to absolute.